Try to refactor out Uids class.
